Changes after testing on Ubuntu 18.04

+++ install.sh
@@ -8,15 +8,21 @@ git submodule update
 
 # Installing powerline fonts for vim
 FONTS=$HOME/fonts
-git clone https://github.com/powerline/fonts.git $FONTS
-bash $FONTS/install.sh
-rm -rf $FONTS
+if ! (fc-list | grep -qi powerline); then
+  git clone https://github.com/powerline/fonts.git $FONTS
+  bash $FONTS/install.sh
+  rm -rf $FONTS
+fi
 
-# Install baseline apps
+echo
+echo "Installing baseline apps"
 sudo apt install\
+  curl\
+  vim\
   tmux\
   i3\
   python3\
+  python3-setuptools\
   compton\
   htop\
   pavucontrol\
@@ -24,33 +30,46 @@ sudo apt install\
   xbacklight\
   stow\
   blueman\
-  feh
+  feh\
+  exfat-fuse\
+  exfat-utils
 
-# Alacritty install
-sudo add-apt-repository ppa:mmstick76/alacritty
+echo
+echo "Adding package repos"
+declare -a ppa_list=("mmstick76/alacritty"\
+                     "js-reynaud/kicad-5.1"\
+                     "team-gcc-arm-embedded/ppa")
+
+for the_ppa in "${ppa_list[@]}"
+do
+  if ! grep -q "$the_ppa" /etc/apt/sources.list /etc/apt/sources.list.d/*; then
+    sudo add-apt-repository ppa:$the_ppa
+  fi
+done
+
+# Chrome is annoying for packages
+if ! grep -q "dl.google.com/linux/chrome/deb/" /etc/apt/sources.list.d/*; then
+  sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
+  wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
+fi
+
 sudo apt-get update
-sudo apt install alacritty
 
-# Kicad install
-sudo add-apt-repository --yes ppa:js-reynaud/kicad-5.1
-sudo apt update
-sudo apt install --install-suggests kicad
+sudo apt install\
+  alacritty\
+  gcc make\
+  kicad\
+  gcc-arm-none-eabi openocd\
+  google-chrome-stable
 
-# Arm gcc install
-sudo add-apt-repository ppa:team-gcc-arm-embedded/ppa
-sudo apt-get update
-sudo apt-get install gcc-arm-none-eabi openocd
+echo
+echo "Installing pip"
+curl https://bootstrap.pypa.io/get-pip.py -o $HOME/get-pip.py
+python3 $HOME/get-pip.py --user
+rm $HOME/get-pip.py
 
-# Chrome install
-sudo sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list'
-wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | sudo apt-key add -
-sudo apt-get update
-sudo apt-get install google-chrome-stable
-
-curl https://bootstrap.pypa.io/get-pip.py -o /tmp/get-pip.py
-python3 /tmp/get-pip.py
-
-# Symbolically link all the dot files!
+echo
+echo "Linking dot files"
 cd $DOTFILES
 stow -v vim\
         bash\
